Create render_gl.go and system_glfw.go files
This makes it easier to have other rendering and windowing backends.
