+++ src/system.go
@@ -357,102 +357,13 @@ type System struct {
 	windowCentered  bool
 }
 
-type Window struct {
-	*glfw.Window
-	title      string
-	fullscreen bool
-	x, y, w, h int
-}
-
-func (s *System) newWindow(w, h int) (*Window, error) {
-	var err error
-	var window *glfw.Window
-	var monitor *glfw.Monitor
-
-	if monitor = glfw.GetPrimaryMonitor(); monitor == nil {
-		return nil, fmt.Errorf("failed to obtain primary monitor")
-	}
-
-	var mode = monitor.GetVideoMode()
-	var x, y = (mode.Width - w) / 2, (mode.Height - h) / 2
-
-	// "-windowed" overrides the configuration setting but does not change it
-	_, forceWindowed := sys.cmdFlags["-windowed"]
-	fullscreen := s.fullscreen && !forceWindowed
-
-	// Create main window.
-	// NOTE: Borderless fullscreen is in reality just a window without borders.
-	if fullscreen && !s.borderless {
-		window, err = glfw.CreateWindow(w, h, s.windowTitle, monitor, nil)
-	} else {
-		window, err = glfw.CreateWindow(w, h, s.windowTitle, nil, nil)
-	}
-	if err != nil {
-		return nil, fmt.Errorf("failed to create window: %w", err)
-	}
-
-	// Set windows atributes
-	if fullscreen {
-		window.SetPos(0, 0)
-		if s.borderless {
-			window.SetAttrib(glfw.Decorated, 0)
-			window.SetSize(mode.Width, mode.Height)
-		}
-		window.SetInputMode(glfw.CursorMode, glfw.CursorHidden)
-	} else {
-		window.SetSize(w, h)
-		window.SetInputMode(glfw.CursorMode, glfw.CursorNormal)
-		if s.windowCentered {
-			window.SetPos(x, y)
-		}
-	}
-
-	window.MakeContextCurrent()
-	window.SetKeyCallback(keyCallback)
-	window.SetCharModsCallback(charCallback)
-	ret := &Window{window, s.windowTitle, fullscreen, x, y, w, h}
-	return ret, err
-}
-
-func (w *Window) toggleFullscreen() {
-	var mode = glfw.GetPrimaryMonitor().GetVideoMode()
-
-	if w.fullscreen {
-		w.SetAttrib(glfw.Decorated, 1)
-		w.SetMonitor(&glfw.Monitor{}, w.x, w.y, w.w, w.h, mode.RefreshRate)
-		w.SetInputMode(glfw.CursorMode, glfw.CursorNormal)
-	} else {
-		w.SetAttrib(glfw.Decorated, 0)
-		if sys.borderless {
-			w.SetSize(mode.Width, mode.Height)
-			w.SetMonitor(&glfw.Monitor{}, 0, 0, mode.Width, mode.Height, mode.RefreshRate)
-		} else {
-			w.x, w.y = w.GetPos()
-			w.SetMonitor(glfw.GetPrimaryMonitor(), w.x, w.y, w.w, w.h, mode.RefreshRate)
-		}
-		w.SetInputMode(glfw.CursorMode, glfw.CursorHidden)
-	}
-	if sys.vRetrace != -1 {
-		glfw.SwapInterval(sys.vRetrace)
-	}
-	w.fullscreen = !w.fullscreen
-}
-
 // Initialize stuff, this is called after the config int at main.go
 func (s *System) init(w, h int32) *lua.LState {
 	s.setWindowSize(w, h)
 	var err error
-	// Create a GLFW window.
-	glfw.WindowHint(glfw.Resizable, glfw.False)
-	glfw.WindowHint(glfw.ContextVersionMajor, 2)
-	glfw.WindowHint(glfw.ContextVersionMinor, 1)
+	// Create a system window.
 	s.window, err = s.newWindow(int(s.scrrect[2]), int(s.scrrect[3]))
 	chk(err)
-
-	// V-Sync
-	if s.vRetrace >= 0 {
-		glfw.SwapInterval(s.vRetrace)
-	}
 
 	// Check if the shader selected is currently available.
 	if s.postProcessingShader < int32(len(s.externalShaderList)) {
@@ -565,8 +476,8 @@ func (s *System) eventUpdate() bool {
 	for _, v := range s.shortcutScripts {
 		v.Activate = false
 	}
-	glfw.PollEvents()
-	s.gameEnd = s.window.Window.ShouldClose()
+	s.window.pollEvents()
+	s.gameEnd = s.window.shouldClose()
 	return !s.gameEnd
 }
 func (s *System) runMainThreadTask() {
